chore(edda): move edda into aws module
