+++ swabbie-aws/src/main/kotlin/com/netflix/spinnaker/swabbie/aws/edda/EddaService.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.swabbie.edda
+package com.netflix.spinnaker.swabbie.aws.edda
 
 import com.netflix.spinnaker.swabbie.aws.autoscalinggroups.AmazonAutoScalingGroup
 import com.netflix.spinnaker.swabbie.aws.images.AmazonImage
