+++ swabbie-aws/src/main/kotlin/com/netflix/spinnaker/swabbie/aws/edda/providers/EddaLaunchConfigurationProvider.kt
@@ -14,70 +14,69 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.swabbie.edda.providers
+package com.netflix.spinnaker.swabbie.aws.edda.providers
 
 import com.netflix.spectator.api.Registry
 import com.netflix.spinnaker.config.EddaApiClient
 import com.netflix.spinnaker.kork.core.RetrySupport
 import com.netflix.spinnaker.swabbie.Parameters
 import com.netflix.spinnaker.swabbie.ResourceProvider
-import com.netflix.spinnaker.swabbie.aws.loadbalancers.AmazonElasticLoadBalancer
-import com.netflix.spinnaker.swabbie.edda.EddaService
-import com.netflix.spinnaker.swabbie.model.LOAD_BALANCER
+import com.netflix.spinnaker.swabbie.aws.launchconfigurations.AmazonLaunchConfiguration
+import com.netflix.spinnaker.swabbie.aws.edda.EddaService
+import com.netflix.spinnaker.swabbie.model.LAUNCH_CONFIGURATION
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
 import retrofit.RetrofitError
 
 @Component
-open class EddaAmazonElasticLoadBalancerProvider(
-  private val eddaApiClients: List<EddaApiClient>,
+open class EddaLaunchConfigurationProvider(
+  eddaApiClients: List<EddaApiClient>,
   private val retrySupport: RetrySupport,
   private val registry: Registry
-) : ResourceProvider<AmazonElasticLoadBalancer>, EddaApiSupport(eddaApiClients, registry) {
+) : ResourceProvider<AmazonLaunchConfiguration>, EddaApiSupport(eddaApiClients, registry) {
   private val log: Logger = LoggerFactory.getLogger(javaClass)
-
-  override fun getAll(params: Parameters): List<AmazonElasticLoadBalancer>? {
+  override fun getAll(params: Parameters): List<AmazonLaunchConfiguration>? {
     withEddaClient(
       region = params["region"] as String,
       accountId = params["account"] as String,
       environment = params["environment"] as String
     )?.run {
-      return getELBs()
+      return getLaunchConfigurations()
     }
 
     return emptyList()
   }
 
-  override fun getOne(params: Parameters): AmazonElasticLoadBalancer? {
+  override fun getOne(params: Parameters): AmazonLaunchConfiguration? {
     withEddaClient(
       region = params["region"] as String,
       accountId = params["account"] as String,
       environment = params["environment"] as String
     )?.run {
-      return getELB(params["loadBalancerName"] as String)
+      return getLaunchConfiguration(params["launchConfigurationName"] as String)
     }
 
     return null
   }
 
-  private fun EddaService.getELBs(): List<AmazonElasticLoadBalancer> {
+  private fun EddaService.getLaunchConfigurations(): List<AmazonLaunchConfiguration> {
     return try {
       retrySupport.retry({
-        this.getLoadBalancers()
+        this.getLaunchConfigs()
       }, maxRetries, retryBackOffMillis, true)
     } catch (e: Exception) {
-      registry.counter(eddaFailureCountId.withTags("resourceType", LOAD_BALANCER)).increment()
-      log.error("failed to get load balancers", e)
+      registry.counter(eddaFailureCountId.withTags("resourceType", LAUNCH_CONFIGURATION)).increment()
+      log.error("failed to get instances", e)
       throw e
     }
   }
 
-  private fun EddaService.getELB(loadBalancerName: String): AmazonElasticLoadBalancer? {
+  private fun EddaService.getLaunchConfiguration(launchConfigurationName: String): AmazonLaunchConfiguration? {
     return try {
       retrySupport.retry({
         try {
-          this.getLoadBalancer(loadBalancerName)
+          this.getLaunchConfig(launchConfigurationName)
         } catch (e: Exception) {
           if (e is RetrofitError && e.response.status == 404) {
             null
@@ -87,8 +86,8 @@ open class EddaAmazonElasticLoadBalancerProvider(
         }
       }, maxRetries, retryBackOffMillis, false)
     } catch (e: Exception) {
-      registry.counter(eddaFailureCountId.withTags("resourceType", LOAD_BALANCER)).increment()
-      log.error("failed to get load balancer {}", loadBalancerName, e)
+      registry.counter(eddaFailureCountId.withTags("resourceType", LAUNCH_CONFIGURATION)).increment()
+      log.error("failed to get launch config {}", launchConfigurationName, e)
       throw e
     }
   }
